AIDD_Task_Day_04/requirements.txt
streamlit>=1.28.0
pypdf>=3.17.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0

